Slight refactor to split tests and consolidate shell commands
