+++ package.json
@@ -5,9 +5,9 @@
   "main": "index.js",
   "repository": "https://github.com/ffffranklin/scotty.git",
   "scripts": {
-    "test": "tape test/**/*.js | faucet",
-    "cover": "istanbul cover tape -- test/**/*.spec.js",
-    "codeclimate": "codeclimate-test-reporter < coverage/lcov.info"
+    "test": "gulp test",
+    "cover": "gulp cover",
+    "codeclimate": "gulp codeclimate"
   },
   "keywords": [
     "synology",
